tetris_a/requirements.txt
asgiref
certifi
clipboard
Django
image
keyboard
MouseInfo
mss
numpy
opencv-python
Pillow
PyAutoGUI
pygame
PyGetWindow
PyMsgBox
pyobjc
pyperclip
PyRect
PyScreeze
pytesseract
pytz
rubicon-objc
six
sqlparse
termcolor
